// sourcecode/BTrak.Packages/snova-audits-module/projects/project-components/src/lib/audits/dependencies/models/asset.ts
export class Assets {
    assetId: string;
    assetNumber: string;
    assetName: string;
    productId: string;
    productName: string;
    productDetailsId: string;
    purchasedDate: Date;
    assignedDateFrom: Date;
    approvedByUserId: string;
    approvedByEmployeeName: string;
    assignedToEmployeeId: string;
    assignedToUserId: string;
    assignedToEmployeeName: string;
    assignedToEmployeeProfileImage: string;
    isSelfApproved: boolean;
    cost: number;
    currencyId: string;
    currencyType: string;
    currencySymbol: string;
    isVendor: boolean;
    isEmpty: boolean;
    isWriteOff: boolean;
    damagedByUserId: string;
    damagedByFullName: string;
    damagedByProfileImage: string;
    damagedDate: Date;
    damagedReason: string;
    manufactureCode: string;
    supplierId: string;
    searchText: string;
    totalCount: number;
    timeStamp: any;
    branchId: string;
    branchName: string;
    seatingId: string;
    seatCode: string;
    assetUniqueNumber: string;
    assetUniqueNumberId: string;
    failedCount: number;
    activeAssignee: boolean;
}